fix: change district parameter type from byte to int in Institute constructor

+++ EduRankCR.Domain/InstituteAggregate/Entities/Institute.cs
@@ -64,7 +64,7 @@ public sealed class Institute : Entity<InstituteId>
         string name,
         byte type,
         byte province,
-        byte district,
+        int district,
         string? url,
         byte status)
     {
